refactor(connector-tester): move SnapshotComparer into Snapshot folder
